nginx php optimization

+++ app/php/include/php/ext/mbstring/libmbfl/mbfl/mbfl_allocators.h
@@ -31,16 +31,15 @@
 #ifndef MBFL_ALLOCATORS_H
 #define MBFL_ALLOCATORS_H
 
-#include <stddef.h>
 #include "mbfl_defs.h"
 
 typedef struct _mbfl_allocators {
-	void *(*malloc)(size_t);
-	void *(*realloc)(void *, size_t);
-	void *(*calloc)(size_t, size_t);
+	void *(*malloc)(unsigned int);
+	void *(*realloc)(void *, unsigned int);
+	void *(*calloc)(unsigned int, unsigned int);
 	void (*free)(void *);
-	void *(*pmalloc)(size_t);
-	void *(*prealloc)(void *, size_t);
+	void *(*pmalloc)(unsigned int);
+	void *(*prealloc)(void *, unsigned int);
 	void (*pfree)(void *);
 } mbfl_allocators;
